Send User-Agent as a custom header (#157)
Setting `User-Agent` request header gives us a steady stream of errors
in Chromium browsers, "protected header" or something.  By using our own
arbitrary `X-App-Identifier` header instead, we can pass the app
identifier without browser complaints. On the other hand, a server
monitoring tool will not show this information to the server operator,
unless they inspect the details on a individual request... Still better
than before! :)

(Also removing the "Content-Type" header, as it does not apply to GET
requests, which is the only type of requests that we perform this far.)